[CI] 릴리스 SemVer 검증 중복 제거
