+++ tools/release/build-monorepo-component-manifests.mjs
@@ -5,6 +5,7 @@ import { existsSync, mkdtempSync, readFileSync, rmSync, symlinkSync, writeFileSy
 import path from "node:path";
 import { fileURLToPath } from "node:url";
 import { validateSchema } from "../ci/lib/json-schema-lite.mjs";
+import { isSemVer } from "./lib/semver.mjs";
 
 const outputFiles = [
   "mobile-component-manifest.json",
@@ -17,37 +18,6 @@ const issueRefPattern = /^AquilaXk\/(easysubway|easysubway-data|easysubway-platf
 
 function fail(message) {
   throw new Error(message);
-}
-
-function isDigits(value) {
-  return value.length > 0 && [...value].every((character) => character >= "0" && character <= "9");
-}
-
-function isSemVerIdentifier(value, rejectNumericLeadingZero) {
-  return value.length > 0
-    && [...value].every((character) => (
-      (character >= "0" && character <= "9")
-      || (character >= "A" && character <= "Z")
-      || (character >= "a" && character <= "z")
-      || character === "-"
-    ))
-    && (!rejectNumericLeadingZero || !isDigits(value) || value === "0" || value[0] !== "0");
-}
-
-function isSemVer(value) {
-  if (typeof value !== "string" || value.length === 0 || value.length > 255) return false;
-  const buildParts = value.split("+");
-  if (buildParts.length > 2) return false;
-  const [version, build] = buildParts;
-  if (build !== undefined && !build.split(".").every((identifier) => isSemVerIdentifier(identifier, false))) return false;
-
-  const prereleaseSeparator = version.indexOf("-");
-  const core = prereleaseSeparator === -1 ? version : version.slice(0, prereleaseSeparator);
-  const prerelease = prereleaseSeparator === -1 ? undefined : version.slice(prereleaseSeparator + 1);
-  if (prerelease !== undefined && !prerelease.split(".").every((identifier) => isSemVerIdentifier(identifier, true))) return false;
-  const coreIdentifiers = core.split(".");
-  return coreIdentifiers.length === 3
-    && coreIdentifiers.every((identifier) => isDigits(identifier) && (identifier === "0" || identifier[0] !== "0"));
 }
 
 function parseArgs(args) {
