+++ tools/release/validate-system-release-manifest.mjs
@@ -4,41 +4,11 @@ import { readFileSync } from "node:fs";
 import path from "node:path";
 import { fileURLToPath } from "node:url";
 import { validateSchema } from "../ci/lib/json-schema-lite.mjs";
+import { isSemVer } from "./lib/semver.mjs";
 
 const slots = ["mobile", "backend", "data", "platform"];
 const canonicalRepositories = Object.fromEntries(slots.map((slot) => [slot, `AquilaXk/easysubway-${slot}`]));
 const allowedRepositories = Object.fromEntries(slots.map((slot) => [slot, new Set(["AquilaXk/easysubway", canonicalRepositories[slot]])]));
-
-function isDigits(value) {
-  return value.length > 0 && [...value].every((character) => character >= "0" && character <= "9");
-}
-
-function isSemVerIdentifier(value, rejectNumericLeadingZero) {
-  return value.length > 0
-    && [...value].every((character) => (
-      (character >= "0" && character <= "9")
-      || (character >= "A" && character <= "Z")
-      || (character >= "a" && character <= "z")
-      || character === "-"
-    ))
-    && (!rejectNumericLeadingZero || !isDigits(value) || value === "0" || value[0] !== "0");
-}
-
-function isSemVer(value) {
-  if (typeof value !== "string" || value.length === 0 || value.length > 255) return false;
-  const buildParts = value.split("+");
-  if (buildParts.length > 2) return false;
-  const [version, build] = buildParts;
-  if (build !== undefined && !build.split(".").every((identifier) => isSemVerIdentifier(identifier, false))) return false;
-
-  const prereleaseSeparator = version.indexOf("-");
-  const core = prereleaseSeparator === -1 ? version : version.slice(0, prereleaseSeparator);
-  const prerelease = prereleaseSeparator === -1 ? undefined : version.slice(prereleaseSeparator + 1);
-  if (prerelease !== undefined && !prerelease.split(".").every((identifier) => isSemVerIdentifier(identifier, true))) return false;
-  const coreIdentifiers = core.split(".");
-  return coreIdentifiers.length === 3
-    && coreIdentifiers.every((identifier) => isDigits(identifier) && (identifier === "0" || identifier[0] !== "0"));
-}
 
 export function validateSystemReleaseManifest({ manifest, componentSchema, systemSchema, issueRefSchema }) {
   const errors = [...validateSchema(systemSchema, manifest).errors];
